Move all UI functions under UI.Console + update game engine and tests to reflect this
